+++ tests/test_dependencies.py
@@ -19,7 +19,7 @@ def test_run_test_dependency():
     # then
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model1").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model1").execution_airflow_task.downstream_task_ids
     )
     assert (
         task_group_prefix_builder("model1", "run")
@@ -45,7 +45,7 @@ def test_dependency():
 
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model2").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model2").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "run")
@@ -72,11 +72,11 @@ def test_more_complex_dependencies():
     assert tasks.length() == 4
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model2").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model2").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "run")
@@ -88,11 +88,11 @@ def test_more_complex_dependencies():
     )
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model4", "run")
@@ -120,11 +120,11 @@ def test_test_dependencies():
     assert tasks.length() == 4
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model2").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model2").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "run")
@@ -141,7 +141,7 @@ def test_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model1_model2_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model1_model2_test").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model1_model2_test"
@@ -149,7 +149,7 @@ def test_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model1_model2_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model1_model2_test").execution_airflow_task.upstream_task_ids
     )
 
 
@@ -180,7 +180,7 @@ def test_complex_test_dependencies():
     assert tasks.length() == 10
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model2").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model2").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "run")
@@ -188,7 +188,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model3", "run")
@@ -196,11 +196,11 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model4").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model4").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model.dbt_test.model4").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model4").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model4", "run")
@@ -212,7 +212,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model5", "test")
-        in tasks.get_task("model.dbt_test.model7").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model7").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model7", "run")
@@ -220,7 +220,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model6", "test")
-        in tasks.get_task("model.dbt_test.model7").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model7").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model7", "run")
@@ -236,7 +236,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model2_model3_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model2_model3_test").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model2_model3_test"
@@ -244,19 +244,19 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model3", "test")
-        in tasks.get_task("model2_model3_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model2_model3_test").execution_airflow_task.upstream_task_ids
     )
     assert all(
         test_name
         in extract_model_arguments(
-            tasks.get_task("model2_model3_test").run_airflow_task.arguments[0]
+            tasks.get_task("model2_model3_test").execution_airflow_task.arguments[0]
         )
         for test_name in ["test3", "test4", "test5"]
     )
     assert all(
         test_name
         not in extract_model_arguments(
-            tasks.get_task("model2_model3_test").run_airflow_task.arguments[0]
+            tasks.get_task("model2_model3_test").execution_airflow_task.arguments[0]
         )
         for test_name in ["test1", "test2"]
     )
@@ -266,7 +266,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model2", "test")
-        in tasks.get_task("model2_model7_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model2_model7_test").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model2_model7_test"
@@ -274,15 +274,15 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model7", "test")
-        in tasks.get_task("model2_model7_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model2_model7_test").execution_airflow_task.upstream_task_ids
     )
     assert "test2" in extract_model_arguments(
-        tasks.get_task("model2_model7_test").run_airflow_task.arguments[0]
+        tasks.get_task("model2_model7_test").execution_airflow_task.arguments[0]
     )
     assert all(
         test_name
         not in extract_model_arguments(
-            tasks.get_task("model2_model7_test").run_airflow_task.arguments[0]
+            tasks.get_task("model2_model7_test").execution_airflow_task.arguments[0]
         )
         for test_name in ["test1", "test3", "test4", "test5"]
     )
@@ -292,7 +292,7 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model5", "test")
-        in tasks.get_task("model5_model6_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model5_model6_test").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model5_model6_test"
@@ -300,15 +300,15 @@ def test_complex_test_dependencies():
     )
     assert (
         task_group_prefix_builder("model6", "test")
-        in tasks.get_task("model5_model6_test").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model5_model6_test").execution_airflow_task.upstream_task_ids
     )
     assert "test1" in extract_model_arguments(
-        tasks.get_task("model5_model6_test").run_airflow_task.arguments[0]
+        tasks.get_task("model5_model6_test").execution_airflow_task.arguments[0]
     )
     assert all(
         test_name
         not in extract_model_arguments(
-            tasks.get_task("model5_model6_test").run_airflow_task.arguments[0]
+            tasks.get_task("model5_model6_test").execution_airflow_task.arguments[0]
         )
         for test_name in ["test2", "test3", "test4", "test5"]
     )