+++ tests/test_ephemeral_operator.py
@@ -73,14 +73,12 @@ def test_no_ephemeral_dag_factory():
 def test_ephemeral_tasks():
     with test_dag():
         factory = AirflowDagFactory(path.dirname(path.abspath(__file__)), "ephemeral_operator")
-        tasks = factory._builder.parse_manifest_into_tasks(
-            factory._manifest_file_path(factory.read_config())
-        )
+        tasks = factory._builder.parse_manifest_into_tasks(factory._manifest_file_path())
 
     # then
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model1").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model1").execution_airflow_task.downstream_task_ids
     )
     assert (
         task_group_prefix_builder("model1", "run")
@@ -89,7 +87,7 @@ def test_ephemeral_tasks():
 
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model2").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model2").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model2__ephemeral"
@@ -98,50 +96,48 @@ def test_ephemeral_tasks():
 
     assert (
         "model2__ephemeral"
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model3__ephemeral"
-        in tasks.get_task("model.dbt_test.model5").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model5").execution_airflow_task.downstream_task_ids
     )
 
     assert (
         "model3__ephemeral"
-        in tasks.get_task("model.dbt_test.model10").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model10").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model9__ephemeral"
-        in tasks.get_task("model.dbt_test.model10").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model10").execution_airflow_task.upstream_task_ids
     )
     assert (
         "model10__ephemeral"
-        in tasks.get_task("model.dbt_test.model3").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model3").execution_airflow_task.downstream_task_ids
     )
     assert (
         "model10__ephemeral"
-        in tasks.get_task("model.dbt_test.model9").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model9").execution_airflow_task.downstream_task_ids
     )
     assert (
         "model11__ephemeral"
-        in tasks.get_task("model.dbt_test.model10").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model10").execution_airflow_task.downstream_task_ids
     )
     assert (
         "model10__ephemeral"
-        in tasks.get_task("model.dbt_test.model11").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model11").execution_airflow_task.upstream_task_ids
     )
 
 
 def test_no_ephemeral_tasks():
     with test_dag():
         factory = AirflowDagFactory(path.dirname(path.abspath(__file__)), "no_ephemeral_operator")
-        tasks = factory._builder.parse_manifest_into_tasks(
-            factory._manifest_file_path(factory.read_config()), show_ephemeral_models=False
-        )
+        tasks = factory._builder.parse_manifest_into_tasks(factory._manifest_file_path())
 
     # then
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model1").run_airflow_task.downstream_task_ids
+        in tasks.get_task("model.dbt_test.model1").execution_airflow_task.downstream_task_ids
     )
     assert (
         task_group_prefix_builder("model1", "run")
@@ -150,7 +146,7 @@ def test_no_ephemeral_tasks():
 
     assert (
         task_group_prefix_builder("model1", "test")
-        in tasks.get_task("model.dbt_test.model4").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model4").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model4", "run")
@@ -159,7 +155,7 @@ def test_no_ephemeral_tasks():
 
     assert (
         task_group_prefix_builder("model6", "test")
-        in tasks.get_task("model.dbt_test.model4").run_airflow_task.upstream_task_ids
+        in tasks.get_task("model.dbt_test.model4").execution_airflow_task.upstream_task_ids
     )
     assert (
         task_group_prefix_builder("model4", "run")